fix(audit): apply all audit findings — migrations, schema, tests, null guards
B1: Migration 0059 now uses expand-migrate-shrink pattern to safely
    rename PENDING→READY_FOR_PACKING with UPDATE for existing rows
W2: Migration 0061 uses TINYINT(1) instead of INT for isPhotographyComplete
    boolean flag; schema.ts updated to use Drizzle boolean() type
W6: Removed silent quoteStatus null fallbacks (|| "UNSENT") in quotes
    router and ordersDb; now throws explicit error on null status
W7: Added 41 new tests across 3 files:
    - quotes.test.ts: status transitions, deprecated values, null handling
    - inventoryMovements.test.ts: adjustment reason enum validation
    - photography.test.ts: isPhotographyComplete flag, enum removal

All ORM calls updated from integer (0/1) to boolean (true/false) for
isPhotographyComplete in catalogPublishingService and photography router.

V4 QA Gate: PASS (check, lint, 5905 tests, build)

+++ drizzle/0061_photography_complete_flag_migration.sql
@@ -1,7 +1,7 @@
 -- TER-574: Photography Complete Flag Migration (enum to boolean)
 --
 -- Step 1: Add isPhotographyComplete boolean column
-ALTER TABLE `batches` ADD COLUMN `isPhotographyComplete` INT NOT NULL DEFAULT 0;
+ALTER TABLE `batches` ADD COLUMN `isPhotographyComplete` TINYINT(1) NOT NULL DEFAULT 0;
 -- Step 2: Migrate data - set flag for batches currently in PHOTOGRAPHY_COMPLETE status
 UPDATE `batches` SET `isPhotographyComplete` = 1 WHERE `batchStatus` = 'PHOTOGRAPHY_COMPLETE';
 -- Step 3: Move PHOTOGRAPHY_COMPLETE batches back to LIVE status
